test/requirements.txt
tonos-ts4==0.4.1
ton-client-py==1.33.1.0
stringcase==1.2.0
decohints==1.0.4
